+++ src/crushinator.js
@@ -4,13 +4,9 @@ Library of simple JS methods to produce crushed image URLs.
 http://github.com/tedconf/js-crushinator-helpers
 */
 
-'use strict';
-
-import * as cropOption from './lib/crop-option';
-import {ParamBuilder} from './lib/param-builder';
-import {prepNumber} from './lib/preppers';
-import {serialize} from './lib/query-string';
-import {warn} from './lib/log';
+import { parameterize } from './lib/parameterize';
+import { serialize } from './lib/query-string';
+import { warn } from './lib/log';
 
 /**
 A list of strings and regular expressions
@@ -18,6 +14,13 @@ A list of strings and regular expressions
 const imageHosts = [
   'assets.tedcdn.com',
   'pb-assets.tedcdn.com',
+  'pa.tedcdn.com',
+  'pe.tedcdn.com',
+  'pf.tedcdn.com',
+  'ph.tedcdn.com',
+  'pj.tedcdn.com',
+  'pk.tedcdn.com',
+  'pl.tedcdn.com',
   'assets2.tedcdn.com',
   'tedcdnpf-a.akamaihd.net',
   'tedcdnpa-a.akamaihd.net',
@@ -37,16 +40,6 @@ const imageHosts = [
 ];
 
 /**
-Possible options parameters for Crushinator.
-*/
-const params = new ParamBuilder({
-  width: { param: 'w', filter: prepNumber },
-  height: { param: 'h', filter: prepNumber },
-  quality: { param: 'quality', filter: prepNumber },
-  crop: cropOption,
-});
-
-/**
 Returns the portion of input URL that corresponds to the host name.
 
 @private
@@ -54,8 +47,15 @@ Returns the portion of input URL that corresponds to the host name.
 @returns {string}
 */
 function extractHost(url) {
-  return url.replace(/.*\/\/([^\/]+).*/, '$1');
+  return String(url).replace(/.*\/\/([^/]+).*/, '$1');
 }
+
+/**
+Overridable global configuration options.
+*/
+export const config = {
+  host: 'https://pi.tedcdn.com',
+};
 
 /**
 Check to see if a URL passes Crushinator's host whitelist.
@@ -74,13 +74,11 @@ Restore a previously crushed URL to its original form.
 @returns {string}
 */
 export function uncrush(url) {
-  const parts = url.match(
-    /(.+)?\/\/(?:img(?:-ssl)?\.tedcdn\.com|tedcdnpi-a\.akamaihd\.net)\/r\/([^?]+)/
-  );
+  const parts = String(url).match(/(.+)?\/\/(?:(?:img(?:-ssl)?|pi)\.tedcdn\.com|tedcdnpi-a\.akamaihd\.net)\/r\/([^?]+)/);
 
   // Avoid double-crushing images
   if (parts) {
-    return uncrush(parts[1] + '//' + parts[2]);
+    return uncrush(`${parts[1]}//${parts[2]}`);
   }
 
   return url;
@@ -91,7 +89,7 @@ Returns a version of the image URL that uses Crushinator with the
 specified options string:
 
     crush('http://images.ted.com/image.jpg', 'w=320')
-      => 'https://tedcdnpi-a.akamaihd.net/images.ted.com/image.jpg?w=320'
+      => 'https://pi.tedcdn.com/images.ted.com/image.jpg?w=320'
 
 @public
 @param {string} url - URL of image to be optimized.
@@ -100,6 +98,11 @@ specified options string:
 @param {number} [options.height] - Target image height in pixels.
 @param {number} [options.quality] - Image quality as a percentage
     (0-100).
+@param {boolean} [options.fit] - Will zoom and crop the image
+    for best fit into the target dimensions.
+@param {string} [options.align] - If cropping occurs, the image
+    can be aligned to the "top", "bottom", "left", "right", or
+    "middle" of the crop frame.
 @param {Object} [options.crop] - Image crop configuration.
 @param {number} [options.crop.width] - Width of cropped section
     in pixels.
@@ -113,32 +116,30 @@ specified options string:
     take place after the image has been resized.
 @returns {string}
 */
-export function crush(url, options={}) {
+export function crush(url, options = {}) {
   // Avoid double-crushing the image
-  url = uncrush(url);
+  const uncrushed = uncrush(url);
 
   // Apply host whitelist
-  if (!crushable(url)) {
-    return url;
+  if (!crushable(uncrushed)) {
+    return uncrushed;
   }
+
+  let params = {};
 
   // Complain about use of the deprecated string API
   if (typeof options === 'string') {
     warn('Sending Crushinator options as a query string is ' +
         'deprecated. Please use the object format.');
+    params = options;
   }
 
   // Stringify object options
-  if (typeof options === 'object') { // or: everything is a duck
-    options = serialize(Object.assign(
-      params.get(options),
-      options.query || {}
-    ));
+  if (typeof options === 'object') {
+    params = serialize(parameterize(options));
   }
 
-  return 'https://tedcdnpi-a.akamaihd.net/r/' +
-    url.replace(/.*\/\//, '') +
-    (options ? '?' + options : '');
+  return `${config.host}/r/${uncrushed.replace(/.*\/\//, '')}${params ? `?${params}` : ''}`;
 }
 
 export default crush;